reverse-proxy: Fix test eval and build
The meaning TLS proxy option has been changed in
e10e9d01ab46de672cb4e849e5882dfd2a509419 and it's no longer just either
"none" or "ACME" but it's now an identifier for nixcloud.TLS.certs.

So for now I'm just removing all of the definitions of proxyOptions.TLS
and set all of the modes to "selfsigned".

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight