reverse-proxy: Refactor folds in checkWebsockets
It's really not necessary to use folds here, because we're basically
working on lists without actually needing an accumulation value.

So all of these folds were basically of the same nature:

fold (x: acc: if cond x then true else acc) false

Let's say we have the following fold:

fold (x: acc: if x == 1 then true else acc) false [ 3 1 2 ]

When unfolding this, we get the following iterations with expanded
values:

(3: false: if 3 == 1 then true else false) -> false
(1: false: if 1 == 1 then true else false) -> true
(2: false: if 2 == 1 then true else false) -> true

The pattern here is similar to builtins.any, but builtins.any is not
only faster because it's a builtin but also has the advantage that it
stops as soon as the first condition becomes true, so it doesn't need to
traverse the whole list.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight