tests/email: Add test for quotas
We're setting a quota of 10 KB for the 'bar' account and we first send a
normal message so that we can make sure the account works in general and
then another message which should exceed the quota.

Note that we're not sending anything to 'bar' from within other tests,
because those tests could be ordered after the quota test and it would
then possibly fail.

Signed-off-by: aszlig <aszlig@nix.build>