tests/email: Make test cases more DRY
There was a lot of repitition for all those tests, so let's actually
create a new helper method wait_for_one_email(), which has that
assertion to check that there is only one new email in the inbox.

In addition to that, the wait_for_new_emails() method now strips and
decodes the text of the new emails by default.

Signed-off-by: aszlig <aszlig@nix.build>