directories: Add enableACLs option
It's a bit unfortunate that POSIX ACL are backwards-compatible in such a
weird way[1][2] that it places the ACL mask value inside the group ID
bits. Some services such as PostgreSQL check these with a normal stat
call and thus fail, because the permissions seem to be too lax for
secure operation. In reality however, they're not too lax because what's
represented by the group ID bits are just the maximum access rights
rather than the effective ones.

So I thought about different ways to solve this and so far I came up
with very unsatisfying results:

  * Automatically revert to UNIX modes whenever no users and groups are
    set. This obviously won't work because one of the reasons the
    "directories" module was introduced in the first place was to
    propagate the permissions to subdirs and subfiles, which would then
    no longer work as expected.
  * Set the mask to "---" if no users and groups are set, which also
    doesn't work because then new files written to that directory by a
    different user than the owner will cause the new owner to have
    correct effective permissions but the propagated permissions will
    then by overridden by the mask, resulting in no permissions at all.
  * Write a kernel patch which deviates from the POSIX standard on a
    per-file basis (using extended attributes). This is something that
    would actually work and should not be very difficult to implement,
    because while posix_acl_equiv_mode() (in fs/posix_acl.c) is an
    exported symbol (thus we might break something if we change the
    declaration), we can hook in our change in posix_acl_update_mode()
    as we have access to the inode struct from there. Obviously this is
    the most invasive way to solve this problem.

After going for a walk I thought I go for the first option, although not
automatically but with the aforementioned enableACLs option. While I
hate to need to set it manually at least it's a possible solution which
at least should fix PostgreSQL startup.

[1]: https://superuser.com/a/642735
[2]: https://serverfault.com/a/352915

Signed-off-by: aszlig <aszlig@nix.build>