release.nix: Add channel tarball
This restructures the release.nix in a way that all jobs except the
channel itself have to succeed for the channel to advance.

The channel itself creates a custom tarball which injects the right Git
revision into the new default.nix, because by default the version is
determined using the .git directory which isn't available in our channel
tarball.

Signed-off-by: aszlig <aszlig@nix.build>