database: Add option to specify additional owners
While this is pretty straightforward with MySQL, which doesn't have the
notion of an owner or even any notion of fine-grained permissions, this
is a bit more involved for PostgreSQL because we only can have *one*
owner for a database.

So we do it by mapping the authentication of a user which is supposed to
own the database to the user actually owning the database.

I needed this while writing the Hydra webservice, which has multiple
users that need to have access to the same database, but it might also
be useful for other purposes.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight