database: Only run postCreate if it's non-empty
This will cause systemd to fail during the post-create service with an
error like this:

  Service lacks both ExecStart= and ExecStop= setting. Refusing.

It will ignore the unit, but having the error in the log is ugly and
confusing, so let's actually not do a postCreate when there is none to
run.

I've also added a few XXX comments to make sure we refactor service
creation in the database module, because right now for both MySQL and
PostgreSQL we do have the same patterns in both files which screams for
deduplication.

Signed-off-by: aszlig <aszlig@nix.build>