Add library function to hash module options
This recursively walks through all the options passed to the function
and returns a SHA256 hash of all the relevant information about the
options, that is it's default values and the type.

The type attributes contain a few lambdas, so we can't simply do a full
toJSON on the whole attribute. Instead we make sure we gather the
submodule options and wrapped types and bring them in a format that's
hashable.

So in the end its usage it's like done in the tests:

(import lib/hash-options {}) options

Right now, this is not yet exposed to modules via a module arg, but once
we have the first usage of it, we might want to do that.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight