run-webservice: Redirect to proxy port directly
If we have a web service that is dependent on what it's base URL is, we
will get into trouble if the port the proxy uses doesn't match the port
we use for the port forwarding to the host.

So let's actually change the port of the proxy server directly and do a
1:1 forwarding of the ports.

This of course also has the downside that if the user changes the port
to something like 25, we will collide with the mail server on the guest.

However this shouldn't happen in real world anyway except if someone
runs run-webservice.sh with root privileges *and* uses port 25 for the
web server port.

Signed-off-by: aszlig <aszlig@nix.build>