web/services: Add basic Hydra service
This still has a few rough edges, for example we need to find out why
the evaluator and queue runner have Restart = "always" instead of
"on-failure" in the original Hydra module.

What's currently missing entirely is handling of GC roots, but that's
easy to add once we are in polishing phase of the service.

The JSON-API of Hydra is quite annoying to use, so we can't just pass a
few curl calls here and there, because it requires a session to be kept
alive. So what I've done instead is use a small Python program which
calls the API and runs a test evaluation and a corresponding build.

Note that I'm using the Nix-builtin derivation function instead of
stdenv.mkDerivation here because otherwise we need a full copy of
<nixpkgs> within the test machine's closure.

Signed-off-by: aszlig <aszlig@nix.build>