Provide a custom error message for VM tests
Some of our tests could be quite problematic if you either have a slow
machine or not enough RAM, so let's add a message to hint the users on
how they can enable the tests.

This is done by wrapping the whole testScript into an eval block and
exit(1) in case there is an error. The reason why I used exit() instead
of another "die" is that we want to make sure we don't run another exit
handler (which for the NixOS test runner would print the message twice,
once in the test runner and the second time through the logger).

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight