hydra: Allow to run with a separate Nix store
Unfortunately this got way messier than what I expected it to be.

So let's start with the evaluator, which is basically a C++ wrapper
which calls a Perl script (which gathers the inputs) and then again the
final evaluator is in C++ again. While this works to some extent the
problem here is that the code uses different ways of interacting with
the Nix store, so there is currently no bullet-proof "Let's just use the
Nix daemon with a different root and only interact with the daemon".

While I was searching for a way to get the evaluator running smoothly I
stumbled upon a bunch of legacy operations in Nix which still use direct
filesystem access instead of going through the store API. So I decided
to fully chroot the evaluator and use bind mounts to bring in the
required store paths along with /etc and others.

Up until that point the evaluator worked fine, at least for a while...

Next up was the queue-runner, which at least isn't a mixture of Perl and
C++, but this time we have a mixture of store_uri and NIX_REMOTE. The
former is for the *destination* store and the latter is for the store we
want to use for building. Obviously the implementation is heavily
tailored for use with Amazon S3, which the official Hydra is using.

Still, even after patching the queue-runner there was still a problem,
mainly that it expected that <nix/corepkgs> is available *within* our
separate store.

I then went on to chroot the queue-runner as well, hoping that I could
simply bind-mount the closures into the separate store and use
"nix-store --register-validity" to make sure it can lookup corepkgs.

Again, this went wrong as well, because now Nix was trying to change the
owner/permissions of the bind-mounted store paths, which doesn't work
because the global Nix store is mounted read-only (at least on most
NixOS systems), which is a good thing but it balked my plan.

By bind-mounting the global NIX_DATA_DIR into the chroot *outside* of
the store however I got the queue-runner working within the chroot.

Now to the bad part: The chroot for the queue-runner as well as the
chroot for the evaluator are based on the same directory. This means
that as soon as systemd is starting up both in parallel, they will
randomly fail to start because of race-conditions during creation of
leading directories for the bind-mounts.

It's not the end of the world, just a bit ugly and as soon as both are
finished fighting they run fine.

The final part to fix was just the Hydra server, which was using
"nix-store --dump" to create a NAR of a particular store path. While
even nix-store supports NIX_REMOTE for changing the store abstraction it
doesn't work for all subcommands and for --dump it does not but instead
tries to access the absolute store path and thus reference the global
store.

I thought I've fixed that by patching Hydra to switch to nix dump-path
instead, which has a --store flag. However it turns out that neither
dump-path nor nix-store --dump really communicates with the Nix daemon
and both try to access the path directly.

For this reason I disabled useSeparateStore by default for now, because
the binary cache won't work for the separate store yet. We need to
possibly fix Nix, so that dump-path uses the daemon to stream a NAR.

Signed-off-by: aszlig <aszlig@nix.build>