lib/call-test: Fix testScript function argument
The testScript attribute may be a function which gets passed a nodes
attribute, so instead of expecting it to be a string, let's wrap the
function arguments and pass it down to the testScript attribute once if
it is a function.

I've removed the fallback to an empty string, because testScript always
needs to be present anyway.

Signed-off-by: aszlig <aszlig@nix.build>