tests: Move generic config defs into call-test.nix
If the user doesn't have sandboxing enabled for Nix, almost all of the
tests will send out requests to letsencryt.org since the recent
introduction of nixcloud.TLS.

In order to avoid this, our test runner now enforces nixcloud.TLS.certs
to be an empty attribute set. The override is with priority 90, so if we
have a test for TLS one day, we can still set the attribute set using
lib.mkForce.

Another thing that's useful for all test is eatmydata.nix, which is
implemented in a way so that it even works with test configurations that
do not contain any database servers, so we can safely add it.

I've removed all the now redundant imports of eatmydata.nix from all the
tests as well.

Signed-off-by: aszlig <aszlig@nix.build>