hydra: Fix creation of GC roots for global store
So far the GC roots directory was created hydra-store-init.service but
the directory also needs to exist if we're not running a separate store,
otherwise evaluation will fail because it can't create a GC root.

Adding another service just started before hydra-init.service will take
care of this. Note that I'm not using the "directories" module, because
that directory needs to be setgid (which can be done using ACLs as well
though, but let's keep it simple for now).

Signed-off-by: aszlig <aszlig@nix.build>