eatmydata: Don't inject in disabled webservices
If we use eatmydata in a web service that is disabled, we get a service
with just an Environment directive and the service then fails to start
with something like this:

Service lacks both ExecStart= and ExecStop= setting. Refusing.

So let's make sure we only inject libeatmydata into web services which
are actually enabled.

Signed-off-by: aszlig <aszlig@nix.build>