mattermost: Refactor config file generation
Retrieving the Mattermost configuration file via import-from-derivation,
apart from being problematic in restricted eval mode has the main
disadvantage that building during evaluation tremendously increases
evaluation time.

Also so far the configuration merge and validation was a two step
endeavour, which now is one single step.

During validation I also added a condition against the Mattermost
version, because only version 4.x segfaults if the validator isn't run
inside the prefix directory (in our case the main store path for
Mattermost) so we only need to work around this in 4.x. Instead of
symlinking the subdirectories of the mattermost derivation, I simply
cd'ed to it.

I compared the old generated configuration file with the new one by
abusing fromJSON/toJSON in Nix to make sure the configuration file is
ordered and piping it into jq to get pretty-printed output. Both the old
and the new configuration files after this refactor resulted in the same
content.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight