Properly fix options and thus building the manual
This reverts commit 431698f54425503b01d07a1a57716a2fd642957e, because
the commit "fixed" the problem by fixing the manual builder, which
wasn't the real issue here.

The real problem is actually at two points:

First of all the default option for the Mattermost config is set to
values that are computed by evaluating other non-static options. So
simply don't set it in the mkOption, as users can still query the
defaults with nixos-option.

Second, and the bigger one: Two submodules of the reverse proxy options
module were bypassing the module system completely and tried to
re-assign values from the top-level config by directly setting it as the
respective default values.

In principle this has the same problem as the Mattermost case, so if we
want to use values based on the configuration we can still use it, but
we'd need to provide a defaultText which is then rendered in the manual.

The submodules for the reverse proxy options are now merged by
declaratively defining the options from the top-level, rather than
trying to do an indirect reference like this:

  { config }:

  let
    foo = { toplevel }: {
      do.something.with = toplevel.config;
    };
  in foo { toplevel = { inherit config; }; }

Even if we don't write proper submodules something like this can still
be simplified to:

  { config }:

  { do.something.with = config; }

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @seitz