tests/container: Disable tmpfs for the Nix store
We would have two possibilities: Either disable sandboxed builds or
disable the tmpfs for the writable store.

I opted for the latter, because the test should match real-world systems
as much as possible, and in NixOS Unstable and the upcoming version
18.09, sandboxed builds are enabled by default.

The reason the test has failed is that sandboxed builds create a chroot
which bind-mounts or hardlinks only the required store paths into the
chroot, depending on whether the source is a file or a directory.

Within the test VM however, the store is using an overlayfs between a
tmpfs and a 9p share. When hardlinks are used, the link is crossing file
system boundaries, causing a failure during sandbox setup.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @seitz