tests/container: Disable Nix sandbox
Follow-up to 860333149730fe8120114fcbbe9083ae7671d779:

Just disabling tmpfs doesn't do the trick, because we still cross file
system boundaries due to the overlayfs.

The reason why I thought this worked was that my test run was using a
wrong version of nixpkgs and thus it succeeded.

So the only option we have here is to disable the Nix sandbox. There is
also another option, which would involve copying the relevant store
paths from the host. That however would be way too much overhead.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @seitz