directories: Allow paths relative to runtimeDir
We only allowed creating directories relative to the state directory so
far, which is the most common case. However if we want to use the
directories module to provide permissions for the runtime directory we
also need to have a way to specify directories for the runtimeDir in web
services.

The first implementation was by adding another option "inRuntimeDir",
however this turned out to be a very bad idea, because we can't use it
for the same subpaths.

For example consider this:

  directories.foo = {
    inRuntimeDir = true;
    owner = "foo";
  };

  directories.foo = {
    owner = "bar";
  };

This will give an error for the conflicting owner definition and even if
we didn't specify an owner the "foo" subdirectory would *only* be
created in runtimeDir but not in stateDir.

So instead of adding another option to the directories submodule, we now
have a new web services option called runtimeDirectories, which is
basically the same as "directories" but for the runtimeDir.

I also added some comments that are useful for people wanting to
implement more directory options.

Signed-off-by: aszlig <aszlig@nix.build>