directories: Check users and groups during eval
If a user or group didn't exist we got a runtime error so far. This is
especially cruicial if we have services that don't directly require a
directory service unit, so errors like this are going to stay in the
dark.

Now we're checking the users and groups specified in the directories
options and if they don't exist an appropriate assertion is thrown.

Of course, this will also throw an assertion if users and groups are
managed imperatively, but I find it highly unlikely that the directories
module is used in conjunction with that. Nevertheless if there really is
a use case for that we can still introduce an option like "checkUsers",
to turn it off on a by-case basis.

Signed-off-by: aszlig <aszlig@nix.build>