web/core: Restrict permissions on runtimeDir
First of all, this removes the mount unit I originally abused to get
leading directories created (we have a directories module now), we
already have a tmpfs in /run so there really is no need to nest.

Second, let's only allow write access to that directory by the
users/groups of the web service and set it sticky, so that those users
aren't able to remove each other's sockets.

In addition to that if the reverse proxy is enabled, the proxy group has
read and execute access to the runtime directory, so it can't create
sockets but only read/write to existing ones.

Signed-off-by: aszlig <aszlig@nix.build>