directories: Allow to run postUpdate commands
So far we had postCreate and postCreateAsRoot, which only run one time
after the directory has been created and not after the fixup of the
permissions is done.

In order to have more complicated permissions like we need to have for
the runtime directories of the web services, it makes sense to allow
arbitrary commands even after the fixup phase.

The alternative to this would be to have NixOS options for setting the
default ACL as well, which would also need a myriad of options to handle
inheritance and other corner cases. For that I think it's simpler to
just allow arbitrary commands instead of making our options more
complicated. After all there might also be a use for postUpdate that is
not tied to just permissions.

Signed-off-by: aszlig <aszlig@nix.build>