web/core: Set default permissions for proxy group
The runtime directory has read permissions for the proxy group, which is
fine so far. However the default permissions are the same as the
directory's main permissions, so files in that directory will be created
with the proxy group only having read access.

What we do now is adding an extra command after the creation/fixup of
the directory, which sets the default permissions accordingly.

Thanks to the new postUpdate and postUpdateAsRoot options, we now can
get rid of the extra service and also set the sticky bit using
postUpdateAsRoot.

Signed-off-by: aszlig <aszlig@nix.build>