Add a new ip2unix core module
This is the NixOS module implementation of a little command line tool I
wrote and published at https://github.com/nixcloud/ip2unix.

It's using the NixOS module system to validate the rules passed to the
program as file. The assertions on these options are somewhat redundant
but it allows us to more quickly bail out on errors during evaluation
time.

So what this module essentially does is providing a small function in
nclib, which can be used to specify arbitrary rules for turning IP
sockets into Unix domain sockets.

Having something like this available is useful when it comes to having
multiple instances, as we don't need to assign ports to various programs
and can just use Unix sockets wherever possible and only use IP sockets
when we *really* need them.

Signed-off-by: aszlig <aszlig@nix.build>