Add RabbitMQ as a web service messaging module
This allows web services to use RabbitMQ by setting
messaging.rabbitmq.enable to true and using config.tools.useRabbitMQ as
a prefix for commands using RabbitMQ. The latter is using our ip2unix
wrapper to wrap an arbitrary command, so other services that would
usually communicate via IP sockets with RabbitMQ are automatically
redirected to the right Unix socket file.

We need a whole range of ip2unix rules to wrap RabbitMQ accordingly,
because the Erlang VM usually communicates via IP sockets and uses its
own port mapping daemon (epmd) as a hub for node interconnection.

The ip2unix command was specifically tested and implemented with
RabbitMQ and thus the Erlang VM in mind, because it provides for a few
challenges like binding to port 0 or various ways to either explicitly
or implicitly bind to sockets.

Signed-off-by: aszlig <aszlig@nix.build>