services/TLS: Fix evaluation of manual
The TLS service tries to compute values for the default values of the
respective submodule options. However during evaluation of the options
from either the manual or nixos-option and the module system evaluates
the default attributes of all options using <name> as the "name"
submodule argument.

This obviously fails to evaluate because there is a function
"hashIdentifierACMEOptions" which in turn tries to access the
submodule's attribute set and for evaluating options that are not
defined it will bail out with an error.

However, the NixOS module system does provide a defaultText attribute
for options, which allows to supply a pre-computed value. This is useful
for something like derivations or for cases like exactly what we have
here with the TLS module.

In addition to providing defaultText attributes, I also added a mkIf for
the config attribute to prevent it from evaluating the whole machinery
when it's not actually needed, because no certs were provided.

Signed-off-by: aszlig <aszlig@nix.build>
Cc: @qknight