core/ip2unix: Coerce direction attribute to string
The direction attribute initially didn't accept a null value, but since
quite some time it is nullable because the field can be optional, which
is also the default and the ip2unix test is not setting the direction.

So in order to correctly display the direction attribute, let's coerce
it to a string, so the null value is displayed as "null" in the
assertion message.

Signed-off-by: aszlig <aszlig@nix.build>