TLS: Move cert options into its own file
This is the first step in untangling the options from the
nixcloud.TLS.certs submodule and it's currently pretty much just moving
things around without much of refactoring.

I did however change a few things, first of all the email option within
the submodule now is of type nullOr x and defaults to null, so we don't
need to generate the value within the submodule from information we only
have from the top-level module. This is to make the submodule more
reusable.

In addition to that I removed a few superflous uses of "toplevel", which
was just capturing the *local* module arguments. Calling something like
this "toplevel" really is just confusing and unnecessary because we
already have the arguments directly available.

The reason why I moved the cert options out into its own file is that we
need to re-use it if we want to make TLS.certs available via webservice
modules. We could do without re-use, if we just make a TLS.certs option
with type of "unspecified", but that would be quite error-prone (and we
wouldn't get nicely generated documentation).

Signed-off-by: aszlig <aszlig@nix.build>