TLS: Deduplicate common attrs and functionality
So far, both the stateDir attribute and the hashIdentifierACMEOptions
function were duplicated in both default.nix and cert-options.nix, so
let's move them into a common.nix.

I also changed the hashIdentifierACMEOptions a bit, first of all by
renaming it to hashACMEConfig, which is what it essentially does and
also by instead of using an identifier name it now directly expects the
to-be-hashed option definitions. This makes the function more general
and thus can be used from cert-options.nix and default.nix without
cert-options.nix needing to have access to the top-level module scope.

Signed-off-by: aszlig <aszlig@nix.build>