Pass TLS.cert option definitions to the top-level
Unfortunately, separating the TLS cert options from its main module
didn't help a lot here, because we can't re-use the certificate options
submodule inside webservices as they conflict with mkMerge, which in
turn nests the option one attribute level deeper and sets something like
the following:

  { _type = "merge"; contents = [ ... ]; }

This doesn't work very well when used directly on submodules, so instead
of re-using the submodule I just used a value-untyped types.attrs for
the webservices option declaration.

One downside of this is that the option declaration will not be merged
in the same way as the top-level option does, but commonly for the web
services abstraction, those options shouldn't need to be merged at all.

Nevertheless in the long term we need to find a proper solution to this,
but for the time being, I'll leave it as is because at least the
attribute keys are merged properly and we do type checking at the
top-level.

Splitting off the options from the main module however is still a good
idea, because it allows us to better distinguish which code affects what
options and to better isolate the submodule from its main module.

The tests for services like Mattermost still do not work because we do
have an override of nixcloud.TLS.certs in lib/call-test.nix, which
passes an empty attribute set to nixcloud.TLS.certs.

Back then when introducing this in 9ea5ec4a96fc6cd88914dfbbe935744e3b32,
I wrote the following:

  If the user doesn't have sandboxing enabled for Nix, almost all of the
  tests will send out requests to letsencryt.org since the recent
  introduction of nixcloud.TLS.

  In order to avoid this, our test runner now enforces
  nixcloud.TLS.certs to be an empty attribute set. The override is with
  priority 90, so if we have a test for TLS one day, we can still set
  the attribute set using lib.mkForce.

So while it's possible to override this from within tests, we need to
*explicitly* state the TLS certs we want to set, but I'm not sure
whether we want to check whether implicit assignment works. This needs a
bit more discussion and it's not the scope of this commit.

Signed-off-by: aszlig <aszlig@nix.build>