Fix eval if no proxyOptions are defined
This is the case if we do have a custom web service that is not tied the
reverse proxy, which for example is the case for our RabbitMQ test.

Unfortunately, proxyOptions is a submodule which has a few options
without a default value, so testing whether a particular option is set
becomes difficult for webservices because the default value for
proxyOptions is {} and we can't query whether there are definitions
existing *inside* the submodule from another module outside
proxyOptions.

So what I'm doing here is just querying whether the proxyOptions option
has been defined *other* that its default definition and *only* then use
the TLS suboption. This will then *only* lead to an error if the reverse
proxy is in use.

Signed-off-by: aszlig <aszlig@nix.build>