hydra: Fix race condition for gcroots init
I haven't been able to reproduce this race condition yet, but for some
reason the "path missing" unit condition didn't fire for the
hydra-init-gcroots service.

Nevertheless I think it's a good idea to always run this service,
because even if the path exists it doesn't hurt to fix up the
permissions of the gcroots dir.

If we really have a race condition when it comes to startup order, we
will notice very soon, because apparently the condition not firing was
the case so far on the Hydra over at https://hydra.nixcloud.io/.

Signed-off-by: aszlig <aszlig@nix.build>