email: Added per-user spam checking and training
This commit adds per-user bayesian spam filtering and training to the
email module. It does this by making spam filtering a two part process
where the first is run as part of a postfix milter and the second is
run by dovecot as part of a sieve script that is run before any other
sieve scripts. And by training the bayesian filter using the Dovecot
IMAPSieve Plugin so that moving a mail to the Spam folder trains that
mail as spam and moving a mail from Spam folder to any other folder
trains as ham.

This commit also adds an assertion that Rspamd version is at least
1.7.3 since older versions did not support the settings option in the
proxy worker which is used to exclude the rules checked as part of final
delivery from the milter run.

To support the custom rule that loads the spam score from the first run
into the second run I needed Rspamd to support storing the custom Lua
rules in LOCAL_CONFDIR. This change was not added to Rspamd until after
version 1.8.0 and so to support the older versions of Rspamd found in
NixOS 18.09 and NixOS Unstable I have back-ported the change to versions
1.7.3, 1.7.9 and 1.8.0 of Rspamd which are versions I have come across
in NixOS.

The latest version of the Rspamd module in NixOS Unstable include
several changes, made by me, that made it possible to implement this
feature and so to also support NixOS 18.09 this commit also includes a
replacement module for Rspamd that disables the Rspamd module in nixpkgs
and instead downloads and uses a newer one from NixOS Unstable.