call-test: Fix arguments for nixos/lib/testing.nix
In [1], testing.nix now needs an additional "pkgs" argument, which we
now provide on a conditional basis by querying the function args and
only if pkgs is required we actually pass it to testing.nix.

The reason for this is to stay backwards-compatible to older versions of
nixpkgs (for example NixOS 18.09).

[1]: https://github.com/NixOS/nixpkgs/commit/6c68fbd4e1f8beac39cb1f499ff

Signed-off-by: aszlig <aszlig@nix.build>