rabbitmq: Use Python 3.6 for the test client
In nixpkgs-unstable the default Python version is now version 3.7, which
introduces "async" and "await" as reserved keywords.

The "pika" library which we're using however uses "async" as an object
attribute:

  Traceback (most recent call last):
    File ".../bin/.test-rabbitmq-recv-wrapped", line 4, in <module>
      import pika
    File ".../site-packages/pika/__init__.py", line 22, in <module>
      from pika.adapters import BaseConnection
    File ".../site-packages/pika/adapters/__init__.py", line 39, in <module>
      from pika.adapters.libev_connection import LibevConnection
    File ".../site-packages/pika/adapters/libev_connection.py", line 103
      self.async = None
               ^
  SyntaxError: invalid syntax

The current upstream version of pika fixes this error, but it's not yet
in nixpkgs, so until the nixos-unstable channel includes a newer version
of pika, let's use Python 3.6 for now.

Signed-off-by: aszlig <aszlig@nix.build>